+++ src/mnelab/utils/__init__.py
@@ -4,4 +4,10 @@
 
 from mnelab.utils.dependencies import have
 from mnelab.utils.syntax import PythonHighlighter
-from mnelab.utils.utils import Montage, count_locations, image_path, natural_sort
+from mnelab.utils.utils import (
+    Montage,
+    calculate_channel_stats,
+    count_locations,
+    image_path,
+    natural_sort,
+)